+++ controllers/dashboard-routes.js
@@ -19,19 +19,35 @@ router.get("/", (req, res) => {
 
 router.get("/dashboard", withGuard, async (req, res) => {
   try {
-    res.render("teams");
-    // console.log(`Hit the dashboard`);
+    const teamData = await Team.findAll({
+      include: [{ model: Match, through: TeamMatch }],
+    });
+    const teams = teamData.map((team) => team.get({ plain: true }));
+    res.render("dashboard", {
+      teams,
+      logged_in: req.session.logged_in,
+    });
   } catch (err) {
     res.status(500).json(err);
   }
 });
 
-router.get("/teams", withGuard, async (req, res) => {
+router.get("/team/:id", withGuard,async (req, res) => {
   try {
-    res.render("teams");
+    const teamData = await Team.findByPk(req.params.team_id, {
+      include: [{ model: Match, through: TeamMatch }],
+    });
+
+    const teamsID = teamData.get({ plain: true });
+    res.render("teams", {
+      teamsID,
+      logged_in: req.session.logged_in,
+    });
   } catch (err) {
     res.status(500).json(err);
   }
 });
+
+
 
 module.exports = router;
